+++ pyvst/vstplugin.py
@@ -1,9 +1,11 @@
+import contextlib
 from ctypes import (cdll, Structure, POINTER, CFUNCTYPE,
                     c_void_p, c_int, c_float, c_int32, c_double, c_char,
                     addressof, byref, pointer, cast, string_at, create_string_buffer)
 from warnings import warn
 
 import numpy
+from wurlitzer import pipes
 
 from .vstwrap import (
     AudioMasterOpcodes,
@@ -30,13 +32,22 @@ def _default_audio_master_callback(effect, opcode, *args):
 
 
 class VstPlugin:
-    def __init__(self, filename, audio_master_callback=None):
+    def __init__(self, filename, audio_master_callback=None, verbose=False):
+        """
+        :param verbose: Set to True to show the plugin's stdout/stderr. By default (False),
+            we capture it.
+        """
+        self.verbose = verbose
+
         if audio_master_callback is None:
             audio_master_callback = _default_audio_master_callback
         self._lib = cdll.LoadLibrary(filename)
         self._lib.VSTPluginMain.argtypes = [AUDIO_MASTER_CALLBACK_TYPE]
         self._lib.VSTPluginMain.restype = POINTER(AEffect)
-        self._effect = self._lib.VSTPluginMain(AUDIO_MASTER_CALLBACK_TYPE(audio_master_callback)).contents
+
+        # Capture stdout and stderr, unless in verbose mode
+        with pipes() if not verbose else contextlib.suppress():
+            self._effect = self._lib.VSTPluginMain(AUDIO_MASTER_CALLBACK_TYPE(audio_master_callback)).contents
 
         assert self._effect.magic == MAGIC
 
@@ -59,14 +70,15 @@ class VstPlugin:
         if ptr is None:
             ptr = c_void_p(None)
         # self._effect.dispatcher.argtypes = [POINTER(AEffect), c_int32, c_int32, c_int, c_void_p, c_float]
-        output = self._effect.dispatcher(byref(self._effect), c_int32(opcode), c_int32(index), c_int(value), ptr, c_float(opt))
+        with pipes() if not self.verbose else contextlib.suppress():
+            output = self._effect.dispatcher(byref(self._effect), c_int32(opcode), c_int32(index), c_int(value), ptr, c_float(opt))
         return output
 
     # Parameters
     #
     @property
     def num_params(self):
-        return self._effect.numParams
+        return self._effect.num_params
 
     def _get_param_attr(self, index, opcode):
         # It should be VstStringConstants.kVstMaxParamStrLen == 8 but I've encountered some VST
@@ -98,6 +110,14 @@ class VstPlugin:
     @property
     def vst_version(self):
         return self._dispatch(AEffectOpcodes.effGetVstVersion)
+
+    @property
+    def num_inputs(self):
+        return self._effect.num_inputs
+
+    @property
+    def num_outputs(self):
+        return self._effect.num_outputs
 
     @property
     def num_midi_in(self):
@@ -146,12 +166,13 @@ class VstPlugin:
 
         output = self._make_empty_array(sample_frames, self.num_outputs)
 
-        self._effect.process_replacing(
-            byref(self._effect),
-            input,
-            output,
-            sample_frames
-        )
+        with pipes() if not self.verbose else contextlib.suppress():
+            self._effect.process_replacing(
+                byref(self._effect),
+                input,
+                output,
+                sample_frames
+            )
 
         output = numpy.vstack([numpy.ctypeslib.as_array(output[i], shape=(sample_frames,))
                                for i in range(self.num_outputs)])
@@ -165,13 +186,3 @@ class VstPlugin:
 
     def set_sample_rate(self, sample_rate):
         self._dispatch(AEffectOpcodes.effSetSampleRate, opt=sample_rate)
-
-    #
-    # TODO explicitely implement those, it makes it less confusing
-    def __getattr__(self, attr):
-        """We also try getattr(self._effect, attr) so we don't have to wrap all of those."""
-        try:
-            return getattr(self._effect, attr)
-        except AttributeError:
-            pass
-        raise AttributeError('object VstPlugin has no attribute "{0}" and effect has no attribute "{0}'.format(attr))